Rename actions into routes
